DB Connected

+++ src/main/java/com/walmart/controller/ProductController.java
@@ -21,7 +21,7 @@ public class ProductController {
 	ProductService productService;
 
 	@GetMapping
-	public String getProductDetailsById(int productId) {
+	public String getProductDetailsById(Long productId) {
 
 		String productDescription;
