+++ src/main/java/com/walmart/service/ProductService.java
@@ -2,6 +2,6 @@ package com.walmart.service;
 
 public interface ProductService {
 	
-	public String getProductDescription(int productId);
+	public String getProductDescription(Long productId);
 
 }
